Build with -static-libgcc for 32-bit Windows
It seems the default is -shared-libgcc, as the 5.3 and 5.4 DLLs get an
entry in the import table stating their dependency on
libgcc_s_dw2-1.dll. The only symbols they import are integer signed
division/modulus subroutines, so dynamic linking looks out of place
here.

Strangely enough, the 5.2 library doesn't have that dependency.